refactor: move delta utils to a specif folder in utils
